+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+3.2.2
+=====
+
+* (improvement) Add helper `ReleaseVersion::createFromPreviewFlag()`.
+* (feature) Add `ComponentDataVisitorInterface` to also be able to visit components.
+
+
 3.2.1
 =====